refactor(Base.mjs): rename module for the class Base
